+++ src/main/catalog-sources/__test__/kubeconfig-sync.test.ts
@@ -2,7 +2,7 @@ import { ObservableMap } from "mobx";
 import { CatalogEntity } from "../../../common/catalog";
 import { loadFromOptions } from "../../../common/kube-helpers";
 import { Cluster } from "../../cluster";
-import { computeDiff, configToModels } from "../kubeconfig-sync.source";
+import { computeDiff, configToModels } from "../kubeconfig-sync";
 import mockFs from "mock-fs";
 import fs from "fs";
