+++ src/renderer/bootstrap.tsx
@@ -21,7 +21,7 @@ import { LensApp } from "./lens-app";
 import { ThemeStore } from "./theme.store";
 import { HelmRepoManager } from "../main/helm/helm-repo-manager";
 import { ExtensionInstallationStateStore } from "./components/+extensions/extension-install.store";
-import { createMuiTheme, ThemeProvider } from "@material-ui/core";
+import { DefaultProps } from "./mui-base-theme";
 
 /**
  * If this is a development buid, wait a second to attach
@@ -95,38 +95,6 @@ export async function bootstrap(App: AppComponent) {
     {isMac && <div id="draggable-top" />}
     {DefaultProps(App)}
   </>, rootElem);
-}
-
-const defaultTheme = createMuiTheme({
-  props: {
-    MuiIconButton: {
-      color: "inherit",
-    },
-    MuiSvgIcon: {
-      fontSize: "inherit",
-    },
-    MuiTooltip: {
-      placement: "top",
-    }
-  },
-  overrides: {
-    MuiIconButton: {
-      root: {
-        "&:hover": {
-          color: "var(--iconActiveColor)",
-          backgroundColor: "var(--iconActiveBackground)",
-        }
-      }
-    }
-  },
-});
-
-function DefaultProps(App: AppComponent) {
-  return (
-    <ThemeProvider theme={defaultTheme}>
-      <App />
-    </ThemeProvider>
-  );
 }
 
 // run
